Move journeys, fixtures and pages into appropriate directories
