+++ e2e/fixtures/citizen/activateCase/applicantDashboardContent.ts
