+++ e2e/fixtures/citizen/activateCase/caseActivatedContent.ts
