+++ e2e/fixtures/citizen/activateCase/enterPinPageContent.ts
