+++ e2e/fixtures/citizen/activateCase/respondentDashboardContent.ts
